services/order-service/requirements.txt
Django>=5.2
djangorestframework>=3.15
django-cors-headers>=4.3
psycopg[binary]>=3.2
requests>=2.32
gunicorn>=22.0

